Try to reduce queries
